+++ deploy/internal/statefulset-core.yaml
@@ -89,6 +89,10 @@ spec:
                 secretKeyRef:
                   name: noobaa-server
                   key: server_secret
+            - name: AGENT_IMAGE
+              valueFrom:
+                fieldRef:
+                  fieldPath: spec.template.spec.containers[1].image
             - name: OAUTH_AUTHORIZATION_ENDPOINT
               valueFrom:
                 configMapKeyRef:
