+++ deploy/olm-catalog/package/noobaa-operator.package.yaml
@@ -1,5 +1,5 @@
 packageName: noobaa-operator
 channels:
 - name: alpha
-  currentCSV: noobaa-operator.v1.0.2
+  currentCSV: noobaa-operator.v1.0.3
 defaultChannel: alpha
