+++ deploy/olm-catalog/package/noobaa-operator.v1.0.1.clusterserviceversion.yaml
@@ -5,7 +5,7 @@ metadata:
     categories: Storage,Big Data
     capabilities: Basic Install
     repository: https://github.com/noobaa/noobaa-operator
-    containerImage: noobaa/noobaa-operator:1.0.2
+    containerImage: noobaa/noobaa-operator:1.0.3
     createdAt: 2019-07-08T13:10:20.940Z
     certified: "false"
     description: NooBaa is an object data service for hybrid and multi cloud environments.
@@ -59,11 +59,11 @@ metadata:
           }
         }
       ]
-  name: noobaa-operator.v1.0.2
+  name: noobaa-operator.v1.0.3
   namespace: placeholder
 spec:
   displayName: NooBaa Operator
-  version: "1.0.2"
+  version: "1.0.3"
   replaces: noobaa-operator.v0.1.0
   maturity: alpha
   provider:
@@ -170,7 +170,7 @@ spec:
                   valueFrom:
                     fieldRef:
                       fieldPath: metadata.namespace
-                image: noobaa/noobaa-operator:1.0.2
+                image: noobaa/noobaa-operator:1.0.3
                 imagePullPolicy: IfNotPresent
                 name: noobaa-operator
                 resources:
