+++ deploy/operator.yaml
@@ -16,7 +16,7 @@ spec:
       serviceAccountName: noobaa-operator
       containers:
         - name: noobaa-operator
-          image: noobaa/noobaa-operator:1.0.2
+          image: noobaa/noobaa-operator:1.0.3
           imagePullPolicy: IfNotPresent
           resources:
             limits:
