+++ main.go
@@ -10,5 +10,5 @@ import (
 )
 
 func main() {
-	cli.New().Run()
+	cli.Cmd().Execute()
 }
