+++ pkg/util/util.go
@@ -2,9 +2,15 @@ package util
 
 import (
 	"context"
+	"crypto/rand"
+	"encoding/base64"
+	"encoding/hex"
 	"strings"
 
+	"github.com/noobaa/noobaa-operator/pkg/apis"
 	"github.com/sirupsen/logrus"
+	corev1 "k8s.io/api/core/v1"
+	apiextv1beta1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -13,50 +19,68 @@ import (
 	"k8s.io/client-go/discovery"
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
 )
 
 var ctx = context.TODO()
+var lazyConfig *rest.Config
+var lazyRest *rest.RESTClient
+var lazyClient client.Client
+
+func init() {
+	Panic(apiextv1beta1.AddToScheme(scheme.Scheme))
+	Panic(apis.AddToScheme(scheme.Scheme))
+}
 
 // KubeConfig loads kubernetes client config from default locations (flags, user dir, etc)
 func KubeConfig() *rest.Config {
-	config, err := config.GetConfig()
-	Panic(err)
-	return config
+	if lazyConfig == nil {
+		var err error
+		lazyConfig, err = config.GetConfig()
+		Panic(err)
+	}
+	return lazyConfig
 }
 
 // KubeRest returns a configured kubernetes REST client
 func KubeRest() *rest.RESTClient {
-	config := KubeConfig()
-	restClient, err := rest.RESTClientFor(config)
-	Panic(err)
-	return restClient
+	if lazyRest == nil {
+		var err error
+		config := KubeConfig()
+		lazyRest, err = rest.RESTClientFor(config)
+		Panic(err)
+	}
+	return lazyRest
 }
 
 // KubeClient resturns a controller-runtime client
 // We use a lazy mapper and a specialized implementation of fast mapper
 // in order to avoid lags when running a CLI client to a far away cluster.
 func KubeClient() client.Client {
-	config := KubeConfig()
-	mapper := meta.NewLazyRESTMapperLoader(func() (meta.RESTMapper, error) {
-		dc, err := discovery.NewDiscoveryClientForConfig(config)
+	if lazyClient == nil {
+		config := KubeConfig()
+		mapper := meta.NewLazyRESTMapperLoader(func() (meta.RESTMapper, error) {
+			dc, err := discovery.NewDiscoveryClientForConfig(config)
+			Panic(err)
+			return NewFastRESTMapper(dc, func(g *metav1.APIGroup) bool {
+				if g.Name == "" ||
+					g.Name == "apps" ||
+					g.Name == "noobaa.io" ||
+					g.Name == "operator.openshift.io" ||
+					g.Name == "cloudcredential.openshift.io" ||
+					strings.HasSuffix(g.Name, ".k8s.io") {
+					return true
+				}
+				return false
+			}), nil
+		})
+		var err error
+		lazyClient, err = client.New(config, client.Options{Mapper: mapper, Scheme: scheme.Scheme})
 		Panic(err)
-		return NewFastRESTMapper(dc, func(g *metav1.APIGroup) bool {
-			if g.Name == "" ||
-				g.Name == "apps" ||
-				g.Name == "noobaa.io" ||
-				g.Name == "operator.openshift.io" ||
-				g.Name == "cloudcredential.openshift.io" ||
-				strings.HasSuffix(g.Name, ".k8s.io") {
-				return true
-			}
-			return false
-		}), nil
-	})
-	c, err := client.New(config, client.Options{Mapper: mapper, Scheme: scheme.Scheme})
-	Panic(err)
-	return c
+	}
+	return lazyClient
 }
 
 // KubeObject loads a text yaml/json to a kubernets object.
@@ -73,20 +97,21 @@ func KubeObject(text string) runtime.Object {
 
 // KubeApply will check if the object exists and will create/update accordingly
 // and report the object status.
-func KubeApply(c client.Client, obj runtime.Object) bool {
+func KubeApply(obj runtime.Object) bool {
+	klient := KubeClient()
 	objKey, _ := client.ObjectKeyFromObject(obj)
 	gvk := obj.GetObjectKind().GroupVersionKind()
 	clone := obj.DeepCopyObject()
-	err := c.Get(ctx, objKey, clone)
+	err := klient.Get(ctx, objKey, clone)
 	if err == nil {
-		err = c.Update(ctx, obj)
+		err = klient.Update(ctx, obj)
 		if err == nil {
 			logrus.Printf("✅ Updated: %s \"%s\"\n", gvk.Kind, objKey.Name)
 			return false
 		}
 	}
 	if errors.IsNotFound(err) {
-		err = c.Create(ctx, obj)
+		err = klient.Create(ctx, obj)
 		if err == nil {
 			logrus.Printf("✅ Created: %s \"%s\"\n", gvk.Kind, objKey.Name)
 			return true
@@ -102,11 +127,12 @@ func KubeApply(c client.Client, obj runtime.Object) bool {
 
 // KubeCreateSkipExisting will check if the object exists and will create/skip accordingly
 // and report the object status.
-func KubeCreateSkipExisting(c client.Client, obj runtime.Object) bool {
+func KubeCreateSkipExisting(obj runtime.Object) bool {
+	klient := KubeClient()
 	objKey, _ := client.ObjectKeyFromObject(obj)
 	gvk := obj.GetObjectKind().GroupVersionKind()
 	clone := obj.DeepCopyObject()
-	err := c.Get(ctx, objKey, clone)
+	err := klient.Get(ctx, objKey, clone)
 	if err == nil {
 		logrus.Printf("✅ Already Exists: %s \"%s\"\n", gvk.Kind, objKey.Name)
 		return false
@@ -116,7 +142,7 @@ func KubeCreateSkipExisting(c client.Client, obj runtime.Object) bool {
 		return false
 	}
 	if errors.IsNotFound(err) {
-		err = c.Create(ctx, obj)
+		err = klient.Create(ctx, obj)
 		if err == nil {
 			logrus.Printf("✅ Created: %s \"%s\"\n", gvk.Kind, objKey.Name)
 			return true
@@ -140,10 +166,11 @@ func KubeCreateSkipExisting(c client.Client, obj runtime.Object) bool {
 }
 
 // KubeDelete deletes an object and reports the object status.
-func KubeDelete(c client.Client, obj runtime.Object) bool {
+func KubeDelete(obj runtime.Object) bool {
+	klient := KubeClient()
 	objKey, _ := client.ObjectKeyFromObject(obj)
 	gvk := obj.GetObjectKind().GroupVersionKind()
-	err := c.Delete(ctx, obj)
+	err := klient.Delete(ctx, obj)
 	if err == nil {
 		logrus.Printf("❌ Deleted: %s \"%s\"\n", gvk.Kind, objKey.Name)
 		return true
@@ -161,10 +188,11 @@ func KubeDelete(c client.Client, obj runtime.Object) bool {
 }
 
 // KubeCheck checks if the object exists and reports the object status.
-func KubeCheck(c client.Client, obj runtime.Object) bool {
+func KubeCheck(obj runtime.Object) bool {
+	klient := KubeClient()
 	objKey, _ := client.ObjectKeyFromObject(obj)
 	gvk := obj.GetObjectKind().GroupVersionKind()
-	err := c.Get(ctx, objKey, obj)
+	err := klient.Get(ctx, objKey, obj)
 	if err == nil {
 		logrus.Printf("✅ Exists: %s \"%s\"\n", gvk.Kind, objKey.Name)
 		return true
@@ -197,6 +225,95 @@ func Panic(err error) {
 func InitLogger() {
 	logrus.SetLevel(logrus.DebugLevel)
 	logrus.SetFormatter(&logrus.TextFormatter{
-		FullTimestamp: true,
+		// FullTimestamp: true,
 	})
+}
+
+// Logger returns a default logger
+func Logger() *logrus.Entry {
+	return logrus.WithContext(ctx)
+}
+
+// Context returns a default Context
+func Context() context.Context {
+	return ctx
+}
+
+func CurrentNamespace() string {
+	loadingRules := clientcmd.NewDefaultClientConfigLoadingRules()
+	configOverrides := &clientcmd.ConfigOverrides{}
+	kubeConfig := clientcmd.NewNonInteractiveDeferredLoadingClientConfig(loadingRules, configOverrides)
+	ns, _, err := kubeConfig.Namespace()
+	Panic(err)
+	return ns
+}
+
+// PersistentError is an error type that tells the reconcile to avoid requeueing.
+type PersistentError struct {
+	E error
+}
+
+// Error function makes PersistentError implement error interface
+func (e *PersistentError) Error() string { return e.E.Error() }
+
+// assert implement error interface
+var _ error = &PersistentError{}
+
+// NewPersistentError returns a new persistent error.
+func NewPersistentError(err error) *PersistentError {
+	if err == nil {
+		panic("NewPersistentError expects non nil error")
+	}
+	return &PersistentError{E: err}
+}
+
+// IsPersistentError checks if the provided error is persistent.
+func IsPersistentError(err error) bool {
+	_, isPersistent := err.(*PersistentError)
+	return isPersistent
+}
+
+// CombineErrors takes a list of errors and combines them to one.
+// Generally it will return the first non-nil error,
+// but if a persistent error is found it will be returned
+// instead of non-persistent errors.
+func CombineErrors(errs ...error) error {
+	combined := error(nil)
+	for _, err := range errs {
+		if err == nil {
+			continue
+		}
+		if combined == nil {
+			combined = err
+			continue
+		}
+		if IsPersistentError(err) && !IsPersistentError(combined) {
+			combined = err
+		}
+	}
+	return combined
+}
+
+// SecretResetStringDataFromData reads the secret data into string data
+// to streamline the paths that use the secret values as strings.
+func SecretResetStringDataFromData(secret *corev1.Secret) {
+	secret.StringData = map[string]string{}
+	for key, val := range secret.Data {
+		secret.StringData[key] = string(val)
+	}
+	secret.Data = map[string][]byte{}
+}
+
+func RandomBase64(numBytes int) string {
+	randomBytes := make([]byte, numBytes)
+	_, err := rand.Read(randomBytes)
+	Panic(err)
+	return base64.StdEncoding.EncodeToString(randomBytes)
+}
+
+func RandomHex(numBytes int) string {
+	randomBytes := make([]byte, numBytes)
+	_, err := rand.Read(randomBytes)
+	Panic(err)
+	return hex.EncodeToString(randomBytes)
 }
