+++ version/version.go
@@ -2,5 +2,5 @@ package version
 
 const (
 	// Version is the noobaa-operator version (semver)
-	Version = "1.0.2"
+	Version = "1.0.3"
 )
